sys_clone (used for fork and to create threads) now seems to be working.  Not sure how much longer it will take to finish moving all kernel functionality to syscall.  Dynamic libraries, like libc, are no longer emulated.  I'm using 32-bit linux builds for those libraries.

+++ src/wine/builtin/libc/Unistd.java
@@ -319,7 +319,7 @@ public class Unistd {
             thread.setErrno((int)-result);
             return -1;
         }
-        return (int)result;
+        return result;
     }
 
     // int pipe(int fildes[2])
